Change int to float

+++ comp-math/lab2/main-program/internal/methods/methods.go
@@ -26,11 +26,11 @@ type MethodInfo struct {
 	Action func(functions.Function) (root float64, err error)
 }
 
-func HasIntervalRoot(function functions.Function, a int, b int) bool {
+func HasIntervalRoot(function functions.Function, a float64, b float64) bool {
 	return HasIntervalRoots(function, a, b) && function.Derivative(float64(a))*function.Derivative(float64(b)) > 0
 }
 
-func HasIntervalRoots(function functions.Function, a int, b int) bool {
+func HasIntervalRoots(function functions.Function, a float64, b float64) bool {
 	return a*b < 0
 }
